device is ready for proper data collection

+++ device/applications/iotproject/inc/func.h
@@ -13,7 +13,7 @@ void check_modules(void const *argument);
 void murata_process_rx_response(void const *argument);
 void accelerometer_measurement(void);
 void magnetometer_measurement(void);
-void print_accelerometer(double data[]);
+void print_accelerometer(uint16_t data[]);
 void print_magnetometer(uint16_t data[]);
 void Error_Handler(void);
 void IWDG_feed(void const *argument);
